Add requirements.txt
Added pip install requirements for the git. To install the dependencies
type 'pip install -r requirements.txt'

It seems that some of the packages have been installed with a
downloaded file. Testing is required to see if pip will
redownload the files or if it has to be installed
manually.